moved utils from generation to root
