+++ tests/unit/maze_dataset/generation/test_latticemaze.py
@@ -8,7 +8,7 @@ from maze_dataset.generation.lattice_maze import (
     SolvedMaze,
     TargetedLatticeMaze,
 )
-from maze_dataset.generation.utils import adj_list_to_nested_set, bool_array_from_string
+from maze_dataset.utils import adj_list_to_nested_set, bool_array_from_string
 
 
 # thanks to gpt for these tests of _from_pixel_grid
